Move contrib.python to fetch.python (#272)
* Fix lint errors

* Move contrib.python to fetch.python and add to docs
