+++ charmhelpers/fetch/python/version.py
