+++ docs/api/charmhelpers.fetch.python.rst
@@ -1,0 +1,7 @@
+charmhelpers.fetch.python module
+====================================
+
+.. automodule:: charmhelpers.fetch.python
+    :members:
+    :undoc-members:
+    :show-inheritance:
